Show thumbnails as popup
Use the popup mechanism instead of replacing the tab's content. This has
multiple advantages:

1. Fewer permissions needed

We don't need the host_permission and the extension can't modify each
and every YouTube page anymore. We also don't need to acquire the "tabs"
permission, "activeTab" is enough and is even more restricted.

This should be way safer.

2. No scripts running in the background

The javascript for the popup page is only loaded when the popup is
shown. We don't need a background script, and above all, we don't need
to inject a content script into every YouTube tab.

The javascript was small and not very demanding, but still, all these
scripts need to kept running in the background and will also take up
some memory. Not having these scripts at all is even better.

3. Currently loaded page is not lost

Since we don't replace the current page, the page is preserved as well
as the currently playing video etc. You can quickly check the thumbnails
and return to what you were doing without too much disturbance.